# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Elevator Inspection Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  ADMIN: "ADMIN",
  USER: "USER",
} as const;

export interface User {
  id: number;
  email: string;
  role: UserRole;
  isActive: boolean;
  organizationId: number;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  ADMIN: "ADMIN",
  USER: "USER",
} as const;

export interface UpdateUserBody {
  role?: UpdateUserBodyRole;
  isActive?: boolean;
}

export type InviteUserBodyRole =
  (typeof InviteUserBodyRole)[keyof typeof InviteUserBodyRole];

export const InviteUserBodyRole = {
  ADMIN: "ADMIN",
  USER: "USER",
} as const;

export interface InviteUserBody {
  email: string;
  role: InviteUserBodyRole;
  password: string;
}

export interface Customer {
  id: number;
  name: string;
  organizationId: number;
  createdAt: string;
}

export interface CreateCustomerBody {
  name: string;
}

export interface Building {
  id: number;
  name: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  customerId: number;
  customerName?: string;
  organizationId: number;
  createdAt: string;
}

export interface CreateBuildingBody {
  name: string;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  customerId: number;
}

export type ElevatorType = (typeof ElevatorType)[keyof typeof ElevatorType];

export const ElevatorType = {
  traction: "traction",
  hydraulic: "hydraulic",
  other: "other",
} as const;

export interface Elevator {
  id: number;
  name: string;
  description?: string;
  bank?: string;
  type: ElevatorType;
  buildingId: number;
  buildingName?: string;
  customerId?: number;
  customerName?: string;
  organizationId: number;
  createdAt: string;
}

export type CreateElevatorBodyType =
  (typeof CreateElevatorBodyType)[keyof typeof CreateElevatorBodyType];

export const CreateElevatorBodyType = {
  traction: "traction",
  hydraulic: "hydraulic",
  other: "other",
} as const;

export interface CreateElevatorBody {
  name: string;
  description?: string;
  bank?: string;
  type: CreateElevatorBodyType;
  buildingId: number;
}

export type InspectionInspectionType =
  (typeof InspectionInspectionType)[keyof typeof InspectionInspectionType];

export const InspectionInspectionType = {
  CAT1: "CAT1",
  CAT5: "CAT5",
} as const;

export type InspectionStatus =
  (typeof InspectionStatus)[keyof typeof InspectionStatus];

export const InspectionStatus = {
  NOT_STARTED: "NOT_STARTED",
  SCHEDULED: "SCHEDULED",
  IN_PROGRESS: "IN_PROGRESS",
  COMPLETED: "COMPLETED",
  OVERDUE: "OVERDUE",
} as const;

export interface Inspection {
  id: number;
  elevatorId: number;
  elevatorName?: string;
  buildingId?: number;
  buildingName?: string;
  customerId?: number;
  customerName?: string;
  inspectionType: InspectionInspectionType;
  recurrenceYears: number;
  lastInspectionDate?: string;
  nextDueDate?: string;
  scheduledDate?: string;
  completionDate?: string;
  status: InspectionStatus;
  notes?: string;
  organizationId: number;
  createdAt: string;
}

export type CreateInspectionBodyInspectionType =
  (typeof CreateInspectionBodyInspectionType)[keyof typeof CreateInspectionBodyInspectionType];

export const CreateInspectionBodyInspectionType = {
  CAT1: "CAT1",
  CAT5: "CAT5",
} as const;

export type CreateInspectionBodyStatus =
  (typeof CreateInspectionBodyStatus)[keyof typeof CreateInspectionBodyStatus];

export const CreateInspectionBodyStatus = {
  NOT_STARTED: "NOT_STARTED",
  SCHEDULED: "SCHEDULED",
  IN_PROGRESS: "IN_PROGRESS",
  COMPLETED: "COMPLETED",
  OVERDUE: "OVERDUE",
} as const;

export interface CreateInspectionBody {
  elevatorId: number;
  inspectionType: CreateInspectionBodyInspectionType;
  recurrenceYears: number;
  lastInspectionDate?: string;
  scheduledDate?: string;
  completionDate?: string;
  status?: CreateInspectionBodyStatus;
  notes?: string;
}

export interface DashboardSummary {
  totalElevators: number;
  duethisMonth: number;
  overdueCount: number;
  scheduledCount: number;
  totalCustomers: number;
  totalBuildings: number;
}

export interface StatusBreakdownItem {
  status: string;
  count: number;
}

export interface OverdueByBuildingItem {
  buildingId: number;
  buildingName: string;
  customerName?: string;
  overdueCount: number;
}

export type ListCustomersParams = {
  search?: string;
};

export type ListBuildingsParams = {
  customerId?: number;
  search?: string;
};

export type ListElevatorsParams = {
  buildingId?: number;
  customerId?: number;
  type?: ListElevatorsType;
  search?: string;
};

export type ListElevatorsType =
  (typeof ListElevatorsType)[keyof typeof ListElevatorsType];

export const ListElevatorsType = {
  traction: "traction",
  hydraulic: "hydraulic",
  other: "other",
} as const;

export type ListInspectionsParams = {
  elevatorId?: number;
  buildingId?: number;
  customerId?: number;
  status?: ListInspectionsStatus;
  inspectionType?: ListInspectionsInspectionType;
  elevatorType?: ListInspectionsElevatorType;
  search?: string;
  month?: number;
  year?: number;
};

export type ListInspectionsStatus =
  (typeof ListInspectionsStatus)[keyof typeof ListInspectionsStatus];

export const ListInspectionsStatus = {
  NOT_STARTED: "NOT_STARTED",
  SCHEDULED: "SCHEDULED",
  IN_PROGRESS: "IN_PROGRESS",
  COMPLETED: "COMPLETED",
  OVERDUE: "OVERDUE",
} as const;

export type ListInspectionsInspectionType =
  (typeof ListInspectionsInspectionType)[keyof typeof ListInspectionsInspectionType];

export const ListInspectionsInspectionType = {
  CAT1: "CAT1",
  CAT5: "CAT5",
} as const;

export type ListInspectionsElevatorType =
  (typeof ListInspectionsElevatorType)[keyof typeof ListInspectionsElevatorType];

export const ListInspectionsElevatorType = {
  traction: "traction",
  hydraulic: "hydraulic",
  other: "other",
} as const;

export type ExportInspectionsParams = {
  customerId?: number;
  buildingId?: number;
  status?: string;
  inspectionType?: string;
};

export type ExportElevatorsParams = {
  customerId?: number;
  buildingId?: number;
};